feat: abstract parsing logic to tools package
